[REF] hotel: split python files
