+++ hotel/models/__init__.py
@@ -1,3 +1,11 @@
-# See LICENSE file for full copyright and licensing details.
-
-from . import hotel
+from . import account_invoice
+from . import hotel_folio_line
+from . import hotel_folio
+from . import hotel_room_amenities
+from . import hotel_room_type
+from . import hotel_room
+from . import hotel_service_line
+from . import hotel_service_type
+from . import hotel_services
+from . import product_product
+from . import res_company
