--- hotel/models/hotel.py
@@ -1,1601 +1,0 @@
-# See LICENSE file for full copyright and licensing details.
-
-import datetime
-import time
-
-from odoo import _, api, fields, models
-from odoo.exceptions import UserError, ValidationError
-from odoo.osv import expression
-from odoo.tools import DEFAULT_SERVER_DATETIME_FORMAT
-
-
-def _offset_format_timestamp1(
-    src_tstamp_str,
-    src_format,
-    dst_format,
-    ignore_unparsable_time=True,
-    context=None,
-):
-    """
-    Convert a source timeStamp string into a destination timeStamp string,
-    attempting to apply the correct offset if both the server and local
-    timeZone are recognized,or no offset at all if they aren't or if
-    tz_offset is false (i.e. assuming they are both in the same TZ).
-
-    @param src_tstamp_str: the STR value containing the timeStamp.
-    @param src_format: the format to use when parsing the local timeStamp.
-    @param dst_format: the format to use when formatting the resulting
-     timeStamp.
-    @param server_to_client: specify timeZone offset direction (server=src
-                             and client=dest if True, or client=src and
-                             server=dest if False)
-    @param ignore_unparsable_time: if True, return False if src_tstamp_str
-                                   cannot be parsed using src_format or
-                                   formatted using dst_format.
-    @return: destination formatted timestamp, expressed in the destination
-             timezone if possible and if tz_offset is true, or src_tstamp_str
-             if timezone offset could not be determined.
-    """
-    if not src_tstamp_str:
-        return False
-    res = src_tstamp_str
-    if src_format and dst_format:
-        try:
-            # dt_value needs to be a datetime.datetime object\
-            # (so notime.struct_time or mx.DateTime.DateTime here!)
-            dt_value = datetime.datetime.strptime(src_tstamp_str, src_format)
-            if context.get("tz", False):
-                try:
-                    import pytz
-
-                    src_tz = pytz.timezone(context["tz"])
-                    dst_tz = pytz.timezone("UTC")
-                    src_dt = src_tz.localize(dt_value, is_dst=True)
-                    dt_value = src_dt.astimezone(dst_tz)
-                except Exception:
-                    pass
-            res = dt_value.strftime(dst_format)
-        except Exception:
-            # Normal ways to end up here are if strptime or strftime failed
-            if not ignore_unparsable_time:
-                return False
-            pass
-    return res
-
-
-class HotelFloor(models.Model):
-    _name = "hotel.floor"
-    _description = "Floor"
-
-    name = fields.Char("Floor Name", required=True, index=True)
-    sequence = fields.Integer(index=True)
-
-
-class HotelRoomType(models.Model):
-    _name = "hotel.room.type"
-    _description = "Room Type"
-
-    name = fields.Char(required=True)
-    categ_id = fields.Many2one("hotel.room.type", "Category")
-    child_ids = fields.One2many(
-        "hotel.room.type", "categ_id", "Child Categories"
-    )
-
-    @api.multi
-    def name_get(self):
-        def get_names(cat):
-            """ Return the list [cat.name, cat.categ_id.name, ...] """
-            res = []
-            while cat:
-                res.append(cat.name)
-                cat = cat.categ_id
-            return res
-
-        return [(cat.id, " / ".join(reversed(get_names(cat)))) for cat in self]
-
-    @api.model
-    def name_search(self, name, args=None, operator="ilike", limit=100):
-        if not args:
-            args = []
-        if name:
-            # Be sure name_search is symmetric to name_get
-            category_names = name.split(" / ")
-            parents = list(category_names)
-            child = parents.pop()
-            domain = [("name", operator, child)]
-            if parents:
-                names_ids = self.name_search(
-                    " / ".join(parents),
-                    args=args,
-                    operator="ilike",
-                    limit=limit,
-                )
-                category_ids = [name_id[0] for name_id in names_ids]
-                if operator in expression.NEGATIVE_TERM_OPERATORS:
-                    categories = self.search([("id", "not in", category_ids)])
-                    domain = expression.OR(
-                        [[("categ_id", "in", categories.ids)], domain]
-                    )
-                else:
-                    domain = expression.AND(
-                        [[("categ_id", "in", category_ids)], domain]
-                    )
-                for i in range(1, len(category_names)):
-                    domain = [
-                        [
-                            (
-                                "name",
-                                operator,
-                                " / ".join(category_names[-1 - i :]),
-                            )
-                        ],
-                        domain,
-                    ]
-                    if operator in expression.NEGATIVE_TERM_OPERATORS:
-                        domain = expression.AND(domain)
-                    else:
-                        domain = expression.OR(domain)
-            categories = self.search(
-                expression.AND([domain, args]), limit=limit
-            )
-        else:
-            categories = self.search(args, limit=limit)
-        return categories.name_get()
-
-
-class ProductProduct(models.Model):
-    _inherit = "product.product"
-
-    isroom = fields.Boolean("Is Room")
-    iscategid = fields.Boolean("Is Categ")
-    isservice = fields.Boolean("Is Service")
-
-
-class HotelRoomAmenitiesType(models.Model):
-    _name = "hotel.room.amenities.type"
-    _description = "amenities Type"
-
-    name = fields.Char(required=True)
-    amenity_id = fields.Many2one("hotel.room.amenities.type", "Category")
-    child_ids = fields.One2many(
-        "hotel.room.amenities.type", "amenity_id", "Child Categories"
-    )
-
-    @api.multi
-    def name_get(self):
-        def get_names(cat):
-            """ Return the list [cat.name, cat.amenity_id.name, ...] """
-            res = []
-            while cat:
-                res.append(cat.name)
-                cat = cat.amenity_id
-            return res
-
-        return [(cat.id, " / ".join(reversed(get_names(cat)))) for cat in self]
-
-    @api.model
-    def name_search(self, name, args=None, operator="ilike", limit=100):
-        if not args:
-            args = []
-        if name:
-            # Be sure name_search is symetric to name_get
-            category_names = name.split(" / ")
-            parents = list(category_names)
-            child = parents.pop()
-            domain = [("name", operator, child)]
-            if parents:
-                names_ids = self.name_search(
-                    " / ".join(parents),
-                    args=args,
-                    operator="ilike",
-                    limit=limit,
-                )
-                category_ids = [name_id[0] for name_id in names_ids]
-                if operator in expression.NEGATIVE_TERM_OPERATORS:
-                    categories = self.search([("id", "not in", category_ids)])
-                    domain = expression.OR(
-                        [[("amenity_id", "in", categories.ids)], domain]
-                    )
-                else:
-                    domain = expression.AND(
-                        [[("amenity_id", "in", category_ids)], domain]
-                    )
-                for i in range(1, len(category_names)):
-                    domain = [
-                        [
-                            (
-                                "name",
-                                operator,
-                                " / ".join(category_names[-1 - i :]),
-                            )
-                        ],
-                        domain,
-                    ]
-                    if operator in expression.NEGATIVE_TERM_OPERATORS:
-                        domain = expression.AND(domain)
-                    else:
-                        domain = expression.OR(domain)
-            categories = self.search(
-                expression.AND([domain, args]), limit=limit
-            )
-        else:
-            categories = self.search(args, limit=limit)
-        return categories.name_get()
-
-
-class HotelRoomAmenities(models.Model):
-    _name = "hotel.room.amenities"
-    _description = "Room amenities"
-
-    product_id = fields.Many2one(
-        "product.product",
-        "Product Category",
-        required=True,
-        delegate=True,
-        ondelete="cascade",
-    )
-    categ_id = fields.Many2one(
-        "hotel.room.amenities.type", string="Amenities Category", required=True
-    )
-    product_manager = fields.Many2one("res.users", string="Product Manager")
-
-
-class FolioRoomLine(models.Model):
-    _name = "folio.room.line"
-    _description = "Hotel Room Reservation"
-    _rec_name = "room_id"
-
-    room_id = fields.Many2one("hotel.room", "Room id")
-    check_in = fields.Datetime("Check In Date", required=True)
-    check_out = fields.Datetime("Check Out Date", required=True)
-    folio_id = fields.Many2one("hotel.folio", string="Folio Number")
-    status = fields.Selection(string="state", related="folio_id.state")
-
-
-class HotelRoom(models.Model):
-    _name = "hotel.room"
-    _description = "Hotel Room"
-
-    product_id = fields.Many2one(
-        "product.product",
-        "Product_id",
-        required=True,
-        delegate=True,
-        ondelete="cascade",
-    )
-    floor_id = fields.Many2one(
-        "hotel.floor", "Floor No", help="At which floor the room is located."
-    )
-    max_adult = fields.Integer()
-    max_child = fields.Integer()
-    categ_id = fields.Many2one(
-        "hotel.room.type", string="Room Category", required=True
-    )
-    room_amenities = fields.Many2many(
-        "hotel.room.amenities",
-        "temp_tab",
-        "room_amenities",
-        "rcateg_id",
-        help="List of room amenities. ",
-    )
-    status = fields.Selection(
-        [("available", "Available"), ("occupied", "Occupied")],
-        "Status",
-        default="available",
-    )
-    capacity = fields.Integer("Capacity", required=True)
-    room_line_ids = fields.One2many(
-        "folio.room.line", "room_id", string="Room Reservation Line"
-    )
-    product_manager = fields.Many2one("res.users", "Product Manager")
-
-    @api.constrains("capacity")
-    def check_capacity(self):
-        for room in self:
-            if room.capacity <= 0:
-                raise ValidationError(_("Room capacity must be more than 0"))
-
-    @api.onchange("isroom")
-    def isroom_change(self):
-        """
-        Based on isroom, status will be updated.
-        ----------------------------------------
-        @param self: object pointer
-        """
-        if self.isroom is False:
-            self.status = "occupied"
-        if self.isroom is True:
-            self.status = "available"
-
-    @api.multi
-    def write(self, vals):
-        """
-        Overrides orm write method.
-        @param self: The object pointer
-        @param vals: dictionary of fields value.
-        """
-        if "isroom" in vals and vals["isroom"] is False:
-            vals.update({"color": 2, "status": "occupied"})
-        if "isroom" in vals and vals["isroom"] is True:
-            vals.update({"color": 5, "status": "available"})
-        ret_val = super(HotelRoom, self).write(vals)
-        return ret_val
-
-    @api.multi
-    def set_room_status_occupied(self):
-        """
-        This method is used to change the state
-        to occupied of the hotel room.
-        ---------------------------------------
-        @param self: object pointer
-        """
-        return self.write({"isroom": False, "color": 2})
-
-    @api.multi
-    def set_room_status_available(self):
-        """
-        This method is used to change the state
-        to available of the hotel room.
-        ---------------------------------------
-        @param self: object pointer
-        """
-        return self.write({"isroom": True, "color": 5})
-
-
-class HotelFolio(models.Model):
-    @api.multi
-    def name_get(self):
-        res = []
-        disp = ""
-        for rec in self:
-            if rec.order_id:
-                disp = str(rec.name)
-                res.append((rec.id, disp))
-        return res
-
-    @api.model
-    def name_search(self, name="", args=None, operator="ilike", limit=100):
-        if args is None:
-            args = []
-        args += [("name", operator, name)]
-        mids = self.search(args, limit=100)
-        return mids.name_get()
-
-    @api.model
-    def _needaction_count(self, domain=None):
-        """
-         Show a count of draft state folio on the menu badge.
-         @param self: object pointer
-        """
-        return self.search_count([("state", "=", "draft")])
-
-    @api.model
-    def _get_checkin_date(self):
-        if self._context.get("tz"):
-            to_zone = self._context.get("tz")
-        else:
-            to_zone = "UTC"
-        return _offset_format_timestamp1(
-            time.strftime("%Y-%m-%d 12:00:00"),
-            DEFAULT_SERVER_DATETIME_FORMAT,
-            DEFAULT_SERVER_DATETIME_FORMAT,
-            ignore_unparsable_time=True,
-            context={"tz": to_zone},
-        )
-
-    @api.model
-    def _get_checkout_date(self):
-        if self._context.get("tz"):
-            to_zone = self._context.get("tz")
-        else:
-            to_zone = "UTC"
-        tm_delta = datetime.timedelta(days=1)
-        return (
-            datetime.datetime.strptime(
-                _offset_format_timestamp1(
-                    time.strftime("%Y-%m-%d 12:00:00"),
-                    DEFAULT_SERVER_DATETIME_FORMAT,
-                    DEFAULT_SERVER_DATETIME_FORMAT,
-                    ignore_unparsable_time=True,
-                    context={"tz": to_zone},
-                ),
-                "%Y-%m-%d %H:%M:%S",
-            )
-            + tm_delta
-        )
-
-    @api.multi
-    def copy(self, default=None):
-        """
-        @param self: object pointer
-        @param default: dict of default values to be set
-        """
-        return super(HotelFolio, self).copy(default=default)
-
-    _name = "hotel.folio"
-    _description = "hotel folio new"
-    _rec_name = "order_id"
-    _order = "id"
-
-    name = fields.Char(
-        "Folio Number", readonly=True, index=True, default="New"
-    )
-    order_id = fields.Many2one(
-        "sale.order", "Order", delegate=True, required=True, ondelete="cascade"
-    )
-    checkin_date = fields.Datetime(
-        "Check In",
-        required=True,
-        readonly=True,
-        states={"draft": [("readonly", False)]},
-        default=_get_checkin_date,
-    )
-    checkout_date = fields.Datetime(
-        "Check Out",
-        required=True,
-        readonly=True,
-        states={"draft": [("readonly", False)]},
-        default=_get_checkout_date,
-    )
-    room_lines = fields.One2many(
-        "hotel.folio.line",
-        "folio_id",
-        readonly=True,
-        states={"draft": [("readonly", False)], "sent": [("readonly", False)]},
-        help="Hotel room reservation detail.",
-    )
-    service_lines = fields.One2many(
-        "hotel.service.line",
-        "folio_id",
-        readonly=True,
-        states={"draft": [("readonly", False)], "sent": [("readonly", False)]},
-        help="Hotel services details provided to"
-        "Customer and it will included in "
-        "the main Invoice.",
-    )
-    hotel_policy = fields.Selection(
-        [
-            ("prepaid", "On Booking"),
-            ("manual", "On Check In"),
-            ("picking", "On Checkout"),
-        ],
-        "Hotel Policy",
-        default="manual",
-        help="Hotel policy for payment that "
-        "either the guest has to payment at "
-        "booking time or check-in "
-        "check-out time.",
-    )
-    duration = fields.Float(
-        "Duration in Days",
-        help="Number of days which will automatically "
-        "count from the check-in and check-out date. ",
-    )
-    hotel_invoice_id = fields.Many2one(
-        "account.invoice", "Invoice", copy=False
-    )
-    duration_dummy = fields.Float("Duration Dummy")
-
-    @api.constrains("room_lines")
-    def folio_room_lines(self):
-        """
-        This method is used to validate the room_lines.
-        ------------------------------------------------
-        @param self: object pointer
-        @return: raise warning depending on the validation
-        """
-        folio_rooms = []
-        for room in self[0].room_lines:
-            if room.product_id.id in folio_rooms:
-                raise ValidationError(_("You Cannot Take Same Room Twice"))
-            folio_rooms.append(room.product_id.id)
-
-    @api.onchange("checkout_date", "checkin_date")
-    def onchange_dates(self):
-        """
-        This method gives the duration between check in and checkout
-        if customer will leave only for some hour it would be considers
-        as a whole day.If customer will check in checkout for more or equal
-        hours, which configured in company as additional hours than it would
-        be consider as full days
-        --------------------------------------------------------------------
-        @param self: object pointer
-        @return: Duration and checkout_date
-        """
-        configured_addition_hours = 0
-        wid = self.warehouse_id
-        whouse_com_id = wid or wid.company_id
-        if whouse_com_id:
-            configured_addition_hours = wid.company_id.additional_hours
-        myduration = 0
-        chckin = self.checkin_date
-        chckout = self.checkout_date
-        if chckin and chckout:
-            server_dt = DEFAULT_SERVER_DATETIME_FORMAT
-            chkin_dt = datetime.datetime.strptime(chckin, server_dt)
-            chkout_dt = datetime.datetime.strptime(chckout, server_dt)
-            dur = chkout_dt - chkin_dt
-            sec_dur = dur.seconds
-            if (not dur.days and not sec_dur) or (dur.days and not sec_dur):
-                myduration = dur.days
-            else:
-                myduration = dur.days + 1
-            # To calculate additional hours in hotel room as per minutes
-            if configured_addition_hours > 0:
-                additional_hours = abs((dur.seconds / 60) / 60)
-                if additional_hours >= configured_addition_hours:
-                    myduration += 1
-        self.duration = myduration
-        self.duration_dummy = self.duration
-
-    @api.model
-    def create(self, vals, check=True):
-        """
-        Overrides orm create method.
-        @param self: The object pointer
-        @param vals: dictionary of fields value.
-        @return: new record set for hotel folio.
-        """
-        if not "service_lines" and "folio_id" in vals:
-            tmp_room_lines = vals.get("room_lines", [])
-            vals["order_policy"] = vals.get("hotel_policy", "manual")
-            vals.update({"room_lines": []})
-            folio_id = super(HotelFolio, self).create(vals)
-            for line in tmp_room_lines:
-                line[2].update({"folio_id": folio_id})
-            vals.update({"room_lines": tmp_room_lines})
-            folio_id.write(vals)
-        else:
-            if not vals:
-                vals = {}
-            vals["name"] = self.env["ir.sequence"].next_by_code("hotel.folio")
-            vals["duration"] = vals.get("duration", 0.0) or vals.get(
-                "duration_dummy", 0.0
-            )
-            folio_id = super(HotelFolio, self).create(vals)
-            folio_room_line_obj = self.env["folio.room.line"]
-            h_room_obj = self.env["hotel.room"]
-            try:
-                for rec in folio_id:
-                    if not rec.reservation_id:
-                        for room_rec in rec.room_lines:
-                            prod = room_rec.product_id.name
-                            room_obj = h_room_obj.search([("name", "=", prod)])
-                            room_obj.write({"isroom": False})
-                            vals = {
-                                "room_id": room_obj.id,
-                                "check_in": rec.checkin_date,
-                                "check_out": rec.checkout_date,
-                                "folio_id": rec.id,
-                            }
-                            folio_room_line_obj.create(vals)
-            except Exception:
-                for rec in folio_id:
-                    for room_rec in rec.room_lines:
-                        prod = room_rec.product_id.name
-                        room_obj = h_room_obj.search([("name", "=", prod)])
-                        room_obj.write({"isroom": False})
-                        vals = {
-                            "room_id": room_obj.id,
-                            "check_in": rec.checkin_date,
-                            "check_out": rec.checkout_date,
-                            "folio_id": rec.id,
-                        }
-                        folio_room_line_obj.create(vals)
-        return folio_id
-
-    @api.multi
-    def write(self, vals):
-        """
-        Overrides orm write method.
-        @param self: The object pointer
-        @param vals: dictionary of fields value.
-        """
-        product_obj = self.env["product.product"]
-        h_room_obj = self.env["hotel.room"]
-        folio_room_line_obj = self.env["folio.room.line"]
-        room_lst = []
-        room_lst1 = []
-        for rec in self:
-            for res in rec.room_lines:
-                room_lst1.append(res.product_id.id)
-            if vals and vals.get("duration_dummy", False):
-                vals["duration"] = vals.get("duration_dummy", 0.0)
-            else:
-                vals["duration"] = rec.duration
-            for folio_rec in rec.room_lines:
-                room_lst.append(folio_rec.product_id.id)
-            new_rooms = set(room_lst).difference(set(room_lst1))
-            if len(list(new_rooms)) != 0:
-                room_list = product_obj.browse(list(new_rooms))
-                for rm in room_list:
-                    room_obj = h_room_obj.search([("name", "=", rm.name)])
-                    room_obj.write({"isroom": False})
-                    vals = {
-                        "room_id": room_obj.id,
-                        "check_in": rec.checkin_date,
-                        "check_out": rec.checkout_date,
-                        "folio_id": rec.id,
-                    }
-                    folio_room_line_obj.create(vals)
-            if len(list(new_rooms)) == 0:
-                room_list_obj = product_obj.browse(room_lst1)
-                for rom in room_list_obj:
-                    room_obj = h_room_obj.search([("name", "=", rom.name)])
-                    room_obj.write({"isroom": False})
-                    room_vals = {
-                        "room_id": room_obj.id,
-                        "check_in": rec.checkin_date,
-                        "check_out": rec.checkout_date,
-                        "folio_id": rec.id,
-                    }
-                    folio_romline_rec = folio_room_line_obj.search(
-                        [("folio_id", "=", rec.id)]
-                    )
-                    folio_romline_rec.write(room_vals)
-        return super(HotelFolio, self).write(vals)
-
-    @api.onchange("warehouse_id")
-    def onchange_warehouse_id(self):
-        """
-        When you change warehouse it will update the warehouse of
-        the hotel folio as well
-        ----------------------------------------------------------
-        @param self: object pointer
-        """
-        return self.order_id._onchange_warehouse_id()
-
-    @api.onchange("partner_id")
-    def onchange_partner_id(self):
-        """
-        When you change partner_id it will update the partner_invoice_id,
-        partner_shipping_id and pricelist_id of the hotel folio as well
-        ---------------------------------------------------------------
-        @param self: object pointer
-        """
-        if self.partner_id:
-            partner_rec = self.env["res.partner"].browse(self.partner_id.id)
-            order_ids = [folio.order_id.id for folio in self]
-            if not order_ids:
-                self.partner_invoice_id = partner_rec.id
-                self.partner_shipping_id = partner_rec.id
-                self.pricelist_id = partner_rec.property_product_pricelist.id
-                raise (_("Not Any Order For  %s ") % (partner_rec.name))
-            else:
-                self.partner_invoice_id = partner_rec.id
-                self.partner_shipping_id = partner_rec.id
-                self.pricelist_id = partner_rec.property_product_pricelist.id
-
-    @api.multi
-    def action_done(self):
-        self.state = "done"
-
-    @api.multi
-    def action_invoice_create(self, grouped=False, final=False):
-        """
-        @param self: object pointer
-        """
-        room_lst = []
-        invoice_id = self.order_id.action_invoice_create(
-            grouped=False, final=False
-        )
-        for line in self:
-            values = {"invoiced": True, "hotel_invoice_id": invoice_id}
-            line.write(values)
-            for rec in line.room_lines:
-                room_lst.append(rec.product_id)
-            for room in room_lst:
-                room_rec = self.env["hotel.room"].search(
-                    [("name", "=", room.name)]
-                )
-                room_rec.write({"isroom": True})
-        return invoice_id
-
-    @api.multi
-    def action_invoice_cancel(self):
-        """
-        @param self: object pointer
-        """
-        if not self.order_id:
-            raise UserError(_("Order id is not available"))
-        for sale in self:
-            for line in sale.order_line:
-                line.write({"invoiced": "invoiced"})
-        self.state = "invoice_except"
-        return self.order_id.action_invoice_cancel
-
-    @api.multi
-    def action_cancel(self):
-        """
-        @param self: object pointer
-        """
-        if not self.order_id:
-            raise UserError(_("Order id is not available"))
-        for sale in self:
-            for invoice in sale.invoice_ids:
-                invoice.state = "cancel"
-        return self.order_id.action_cancel()
-
-    @api.multi
-    def action_confirm(self):
-        for order in self.order_id:
-            order.state = "sale"
-            if not order.analytic_account_id:
-                for line in order.order_line:
-                    if line.product_id.invoice_policy == "cost":
-                        order._create_analytic_account()
-                        break
-        config_parameter_obj = self.env["ir.config_parameter"]
-        if config_parameter_obj.sudo().get_param("sale.auto_done_setting"):
-            self.order_id.action_done()
-
-    @api.multi
-    def test_state(self, mode):
-        """
-        @param self: object pointer
-        @param mode: state of workflow
-        """
-        write_done_ids = []
-        write_cancel_ids = []
-        if write_done_ids:
-            test_obj = self.env["sale.order.line"].browse(write_done_ids)
-            test_obj.write({"state": "done"})
-        if write_cancel_ids:
-            test_obj = self.env["sale.order.line"].browse(write_cancel_ids)
-            test_obj.write({"state": "cancel"})
-
-    @api.multi
-    def action_cancel_draft(self):
-        """
-        @param self: object pointer
-        """
-        if not len(self._ids):
-            return False
-        query = "select id from sale_order_line \
-        where order_id IN %s and state=%s"
-        self._cr.execute(query, (tuple(self._ids), "cancel"))
-        cr1 = self._cr
-        line_ids = map(lambda x: x[0], cr1.fetchall())
-        self.write({"state": "draft", "invoice_ids": [], "shipped": 0})
-        sale_line_obj = self.env["sale.order.line"].browse(line_ids)
-        sale_line_obj.write(
-            {
-                "invoiced": False,
-                "state": "draft",
-                "invoice_lines": [(6, 0, [])],
-            }
-        )
-        return True
-
-
-class HotelFolioLine(models.Model):
-    @api.multi
-    def copy(self, default=None):
-        """
-        @param self: object pointer
-        @param default: dict of default values to be set
-        """
-        return super(HotelFolioLine, self).copy(default=default)
-
-    @api.model
-    def _get_checkin_date(self):
-        if "checkin" in self._context:
-            return self._context["checkin"]
-        return time.strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-
-    @api.model
-    def _get_checkout_date(self):
-        if "checkout" in self._context:
-            return self._context["checkout"]
-        return time.strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-
-    _name = "hotel.folio.line"
-    _description = "hotel folio1 room line"
-
-    order_line_id = fields.Many2one(
-        "sale.order.line",
-        string="Order Line",
-        required=True,
-        delegate=True,
-        ondelete="cascade",
-    )
-    folio_id = fields.Many2one(
-        "hotel.folio", string="Folio", ondelete="cascade"
-    )
-    checkin_date = fields.Datetime(
-        "Check In", required=True, default=_get_checkin_date
-    )
-    checkout_date = fields.Datetime(
-        "Check Out", required=True, default=_get_checkout_date
-    )
-    is_reserved = fields.Boolean(
-        "Is Reserved",
-        help="True when folio line created from \
-                                 Reservation",
-    )
-
-    @api.model
-    def create(self, vals, check=True):
-        """
-        Overrides orm create method.
-        @param self: The object pointer
-        @param vals: dictionary of fields value.
-        @return: new record set for hotel folio line.
-        """
-        if "folio_id" in vals:
-            folio = self.env["hotel.folio"].browse(vals["folio_id"])
-            vals.update({"order_id": folio.order_id.id})
-        return super(HotelFolioLine, self).create(vals)
-
-    @api.constrains("checkin_date", "checkout_date")
-    def check_dates(self):
-        """
-        This method is used to validate the checkin_date and checkout_date.
-        -------------------------------------------------------------------
-        @param self: object pointer
-        @return: raise warning depending on the validation
-        """
-        if self.checkin_date >= self.checkout_date:
-            raise ValidationError(
-                _(
-                    "Room line Check In Date Should be \
-                less than the Check Out Date!"
-                )
-            )
-        if self.folio_id.date_order and self.checkin_date:
-            if self.checkin_date <= self.folio_id.date_order:
-                raise ValidationError(
-                    _(
-                        "Room line check in date should be \
-                greater than the current date."
-                    )
-                )
-
-    @api.multi
-    def unlink(self):
-        """
-        Overrides orm unlink method.
-        @param self: The object pointer
-        @return: True/False.
-        """
-        sale_line_obj = self.env["sale.order.line"]
-        fr_obj = self.env["folio.room.line"]
-        for line in self:
-            if line.order_line_id:
-                sale_unlink_obj = sale_line_obj.browse([line.order_line_id.id])
-                for rec in sale_unlink_obj:
-                    room_obj = self.env["hotel.room"].search(
-                        [("name", "=", rec.name)]
-                    )
-                    if room_obj.id:
-                        folio_arg = [
-                            ("folio_id", "=", line.folio_id.id),
-                            ("room_id", "=", room_obj.id),
-                        ]
-                        folio_room_line_myobj = fr_obj.search(folio_arg)
-                        if folio_room_line_myobj.id:
-                            folio_room_line_myobj.unlink()
-                            room_obj.write(
-                                {"isroom": True, "status": "available"}
-                            )
-                sale_unlink_obj.unlink()
-        return super(HotelFolioLine, self).unlink()
-
-    def _get_real_price_currency(
-        self, product, rule_id, qty, uom, pricelist_id
-    ):
-        """Retrieve the price before applying the pricelist
-            :param obj product: object of current product record
-            :parem float qty: total quentity of product
-            :param tuple price_and_rule: tuple(price, suitable_rule) coming
-            from pricelist computation
-            :param obj uom: unit of measure of current order line
-            :param integer pricelist_id: pricelist id of sale order"""
-        PricelistItem = self.env["product.pricelist.item"]
-        field_name = "lst_price"
-        currency_id = None
-        product_currency = None
-        if rule_id:
-            pricelist_item = PricelistItem.browse(rule_id)
-            if (
-                pricelist_item.pricelist_id.discount_policy
-                == "without_discount"
-            ):
-                while (
-                    pricelist_item.base == "pricelist"
-                    and pricelist_item.base_pricelist_id
-                    and pricelist_item.base_pricelist_id.discount_policy
-                    == "without_discount"
-                ):
-                    price, rule_id = pricelist_item.base_pricelist_id.with_context(
-                        uom=uom.id
-                    ).get_product_price_rule(
-                        product, qty, self.folio_id.partner_id
-                    )
-                    pricelist_item = PricelistItem.browse(rule_id)
-
-            if pricelist_item.base == "standard_price":
-                field_name = "standard_price"
-            if (
-                pricelist_item.base == "pricelist"
-                and pricelist_item.base_pricelist_id
-            ):
-                field_name = "price"
-                product = product.with_context(
-                    pricelist=pricelist_item.base_pricelist_id.id
-                )
-                product_currency = pricelist_item.base_pricelist_id.currency_id
-            currency_id = pricelist_item.pricelist_id.currency_id
-
-        product_currency = (
-            product_currency
-            or (product.company_id and product.company_id.currency_id)
-            or self.env.user.company_id.currency_id
-        )
-        if not currency_id:
-            currency_id = product_currency
-            cur_factor = 1.0
-        else:
-            if currency_id.id == product_currency.id:
-                cur_factor = 1.0
-            else:
-                cur_factor = currency_id._get_conversion_rate(
-                    product_currency, currency_id
-                )
-
-        product_uom = self.env.context.get("uom") or product.uom_id.id
-        if uom and uom.id != product_uom:
-            # the unit price is in a different uom
-            uom_factor = uom._compute_price(1.0, product.uom_id)
-        else:
-            uom_factor = 1.0
-        return product[field_name] * uom_factor * cur_factor, currency_id.id
-
-    @api.multi
-    def _get_display_price(self, product):
-        # TO DO: move me in master/saas-16 on sale.order
-        if self.folio_id.pricelist_id.discount_policy == "with_discount":
-            return product.with_context(
-                pricelist=self.folio_id.pricelist_id.id
-            ).price
-        product_context = dict(
-            self.env.context,
-            partner_id=self.folio_id.partner_id.id,
-            date=self.folio_id.date_order,
-            uom=self.product_uom.id,
-        )
-        final_price, rule_id = self.folio_id.pricelist_id.with_context(
-            product_context
-        ).get_product_price_rule(
-            self.product_id,
-            self.product_uom_qty or 1.0,
-            self.folio_id.partner_id,
-        )
-        base_price, currency_id = self.with_context(
-            product_context
-        )._get_real_price_currency(
-            product,
-            rule_id,
-            self.product_uom_qty,
-            self.product_uom,
-            self.folio_id.pricelist_id.id,
-        )
-        if currency_id != self.folio_id.pricelist_id.currency_id.id:
-            base_price = (
-                self.env["res.currency"]
-                .browse(currency_id)
-                .with_context(product_context)
-                .compute(base_price, self.folio_id.pricelist_id.currency_id)
-            )
-        # negative discounts (= surcharge) are included in the display price
-        return max(base_price, final_price)
-
-    @api.multi
-    def _compute_tax_id(self):
-        for line in self:
-            fpos = (
-                line.folio_id.fiscal_position_id
-                or line.folio_id.partner_id.property_account_position_id
-            )
-            # If company_id is set, always filter taxes by the company
-            taxes = line.product_id.taxes_id.filtered(
-                lambda r: not line.company_id
-                or r.company_id == line.company_id
-            )
-            line.tax_id = (
-                fpos.map_tax(
-                    taxes, line.product_id, line.folio_id.partner_shipping_id
-                )
-                if fpos
-                else taxes
-            )
-
-    @api.multi
-    @api.onchange("product_id")
-    def product_id_change(self):
-        """
- -        @param self: object pointer
- -        """
-        if not self.product_id:
-            return {"domain": {"product_uom": []}}
-        vals = {}
-        domain = {
-            "product_uom": [
-                ("category_id", "=", self.product_id.uom_id.category_id.id)
-            ]
-        }
-        if not self.product_uom or (
-            self.product_id.uom_id.id != self.product_uom.id
-        ):
-            vals["product_uom"] = self.product_id.uom_id
-        product = self.product_id.with_context(
-            lang=self.folio_id.partner_id.lang,
-            partner=self.folio_id.partner_id.id,
-            quantity=vals.get("product_uom_qty") or self.product_uom_qty,
-            date=self.folio_id.date_order,
-            pricelist=self.folio_id.pricelist_id.id,
-            uom=self.product_uom.id,
-        )
-
-        result = {"domain": domain}
-
-        title = False
-        message = False
-        warning = {}
-        if product.sale_line_warn != "no-message":
-            title = _("Warning for %s") % product.name
-            message = product.sale_line_warn_msg
-            warning["title"] = title
-            warning["message"] = message
-            result = {"warning": warning}
-            if product.sale_line_warn == "block":
-                self.product_id = False
-                return result
-
-        name = product.name_get()[0][1]
-        if product.description_sale:
-            name += "\n" + product.description_sale
-        vals["name"] = name
-
-        self._compute_tax_id()
-
-        if self.folio_id.pricelist_id and self.folio_id.partner_id:
-            vals["price_unit"] = self.env[
-                "account.tax"
-            ]._fix_tax_included_price_company(
-                self._get_display_price(product),
-                product.taxes_id,
-                self.tax_id,
-                self.company_id,
-            )
-        self.update(vals)
-        return result
-
-    @api.onchange("checkin_date", "checkout_date")
-    def on_change_checkout(self):
-        """
-        When you change checkin_date or checkout_date it will checked it
-        and update the qty of hotel folio line
-        -----------------------------------------------------------------
-        @param self: object pointer
-        """
-        configured_addition_hours = 0
-        fwhouse_id = self.folio_id.warehouse_id
-        fwc_id = fwhouse_id or fwhouse_id.company_id
-        if fwc_id:
-            configured_addition_hours = fwhouse_id.company_id.additional_hours
-        myduration = 0
-        if not self.checkin_date:
-            self.checkin_date = time.strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-        if not self.checkout_date:
-            self.checkout_date = time.strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-        chckin = self.checkin_date
-        chckout = self.checkout_date
-        if chckin and chckout:
-            server_dt = DEFAULT_SERVER_DATETIME_FORMAT
-            chkin_dt = datetime.datetime.strptime(chckin, server_dt)
-            chkout_dt = datetime.datetime.strptime(chckout, server_dt)
-            dur = chkout_dt - chkin_dt
-            sec_dur = dur.seconds
-            if (not dur.days and not sec_dur) or (dur.days and not sec_dur):
-                myduration = dur.days
-            else:
-                myduration = dur.days + 1
-            #            To calculate additional hours in hotel room as per minutes
-            if configured_addition_hours > 0:
-                additional_hours = abs((dur.seconds / 60) / 60)
-                if additional_hours >= configured_addition_hours:
-                    myduration += 1
-        self.product_uom_qty = myduration
-        hotel_room_obj = self.env["hotel.room"]
-        hotel_room_ids = hotel_room_obj.search([])
-        avail_prod_ids = []
-        for room in hotel_room_ids:
-            assigned = False
-            for rm_line in room.room_line_ids:
-                if rm_line.status != "cancel":
-                    if (
-                        self.checkin_date
-                        <= rm_line.check_in
-                        <= self.checkout_date
-                    ) or (
-                        self.checkin_date
-                        <= rm_line.check_out
-                        <= self.checkout_date
-                    ):
-                        assigned = True
-                    elif (
-                        rm_line.check_in
-                        <= self.checkin_date
-                        <= rm_line.check_out
-                    ) or (
-                        rm_line.check_in
-                        <= self.checkout_date
-                        <= rm_line.check_out
-                    ):
-                        assigned = True
-            if not assigned:
-                avail_prod_ids.append(room.product_id.id)
-        domain = {"product_id": [("id", "in", avail_prod_ids)]}
-        return {"domain": domain}
-
-    @api.multi
-    def button_confirm(self):
-        """
-        @param self: object pointer
-        """
-        for folio in self:
-            line = folio.order_line_id
-            line.button_confirm()
-        return True
-
-    @api.multi
-    def button_done(self):
-        """
-        @param self: object pointer
-        """
-        lines = [folio_line.order_line_id for folio_line in self]
-        lines.button_done()
-        self.state = "done"
-        return True
-
-    @api.multi
-    def copy_data(self, default=None):
-        """
-        @param self: object pointer
-        @param default: dict of default values to be set
-        """
-        line_id = self.order_line_id.id
-        sale_line_obj = self.env["sale.order.line"].browse(line_id)
-        return sale_line_obj.copy_data(default=default)
-
-
-class HotelServiceLine(models.Model):
-    @api.multi
-    def copy(self, default=None):
-        """
-        @param self: object pointer
-        @param default: dict of default values to be set
-        """
-        return super(HotelServiceLine, self).copy(default=default)
-
-    @api.model
-    def _service_checkin_date(self):
-        if "checkin" in self._context:
-            return self._context["checkin"]
-        return time.strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-
-    @api.model
-    def _service_checkout_date(self):
-        if "checkout" in self._context:
-            return self._context["checkout"]
-        return time.strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-
-    _name = "hotel.service.line"
-    _description = "hotel Service line"
-
-    service_line_id = fields.Many2one(
-        "sale.order.line",
-        "Service Line",
-        required=True,
-        delegate=True,
-        ondelete="cascade",
-    )
-    folio_id = fields.Many2one("hotel.folio", "Folio", ondelete="cascade")
-    ser_checkin_date = fields.Datetime(
-        "From Date", required=True, default=_service_checkin_date
-    )
-    ser_checkout_date = fields.Datetime(
-        "To Date", required=True, default=_service_checkout_date
-    )
-
-    @api.model
-    def create(self, vals, check=True):
-        """
-        Overrides orm create method.
-        @param self: The object pointer
-        @param vals: dictionary of fields value.
-        @return: new record set for hotel service line.
-        """
-        if "folio_id" in vals:
-            folio = self.env["hotel.folio"].browse(vals["folio_id"])
-            vals.update({"order_id": folio.order_id.id})
-        return super(HotelServiceLine, self).create(vals)
-
-    @api.multi
-    def unlink(self):
-        """
-        Overrides orm unlink method.
-        @param self: The object pointer
-        @return: True/False.
-        """
-        s_line_obj = self.env["sale.order.line"]
-        for line in self:
-            if line.service_line_id:
-                sale_unlink_obj = s_line_obj.browse([line.service_line_id.id])
-                sale_unlink_obj.unlink()
-        return super(HotelServiceLine, self).unlink()
-
-    @api.multi
-    def _compute_tax_id(self):
-        for line in self:
-            fpos = (
-                line.folio_id.fiscal_position_id
-                or line.folio_id.partner_id.property_account_position_id
-            )
-            # If company_id is set, always filter taxes by the company
-            taxes = line.product_id.taxes_id.filtered(
-                lambda r: not line.company_id
-                or r.company_id == line.company_id
-            )
-            line.tax_id = (
-                fpos.map_tax(
-                    taxes, line.product_id, line.folio_id.partner_shipping_id
-                )
-                if fpos
-                else taxes
-            )
-
-    def _get_real_price_currency(
-        self, product, rule_id, qty, uom, pricelist_id
-    ):
-        """Retrieve the price before applying the pricelist
-            :param obj product: object of current product record
-            :parem float qty: total quentity of product
-            :param tuple price_and_rule: tuple(price, suitable_rule)
-            coming from pricelist computation
-            :param obj uom: unit of measure of current order line
-            :param integer pricelist_id: pricelist id of sale order"""
-        PricelistItem = self.env["product.pricelist.item"]
-        field_name = "lst_price"
-        currency_id = None
-        product_currency = None
-        if rule_id:
-            pricelist_item = PricelistItem.browse(rule_id)
-            if (
-                pricelist_item.pricelist_id.discount_policy
-                == "without_discount"
-            ):
-                while (
-                    pricelist_item.base == "pricelist"
-                    and pricelist_item.base_pricelist_id
-                    and pricelist_item.base_pricelist_id.discount_policy
-                    == "without_discount"
-                ):
-                    price, rule_id = pricelist_item.base_pricelist_id.with_context(
-                        uom=uom.id
-                    ).get_product_price_rule(
-                        product, qty, self.folio_id.partner_id
-                    )
-                    pricelist_item = PricelistItem.browse(rule_id)
-
-            if pricelist_item.base == "standard_price":
-                field_name = "standard_price"
-            if (
-                pricelist_item.base == "pricelist"
-                and pricelist_item.base_pricelist_id
-            ):
-                field_name = "price"
-                product = product.with_context(
-                    pricelist=pricelist_item.base_pricelist_id.id
-                )
-                product_currency = pricelist_item.base_pricelist_id.currency_id
-            currency_id = pricelist_item.pricelist_id.currency_id
-
-        product_currency = (
-            product_currency
-            or (product.company_id and product.company_id.currency_id)
-            or self.env.user.company_id.currency_id
-        )
-        if not currency_id:
-            currency_id = product_currency
-            cur_factor = 1.0
-        else:
-            if currency_id.id == product_currency.id:
-                cur_factor = 1.0
-            else:
-                cur_factor = currency_id._get_conversion_rate(
-                    product_currency, currency_id
-                )
-
-        product_uom = self.env.context.get("uom") or product.uom_id.id
-        if uom and uom.id != product_uom:
-            # the unit price is in a different uom
-            uom_factor = uom._compute_price(1.0, product.uom_id)
-        else:
-            uom_factor = 1.0
-        return product[field_name] * uom_factor * cur_factor, currency_id.id
-
-    @api.multi
-    def _get_display_price(self, product):
-        # TO DO: move me in master/saas-16 on sale.order
-        if self.folio_id.pricelist_id.discount_policy == "with_discount":
-            return product.with_context(
-                pricelist=self.folio_id.pricelist_id.id
-            ).price
-        product_context = dict(
-            self.env.context,
-            partner_id=self.folio_id.partner_id.id,
-            date=self.folio_id.date_order,
-            uom=self.product_uom.id,
-        )
-        final_price, rule_id = self.folio_id.pricelist_id.with_context(
-            product_context
-        ).get_product_price_rule(
-            self.product_id,
-            self.product_uom_qty or 1.0,
-            self.folio_id.partner_id,
-        )
-        base_price, currency_id = self.with_context(
-            product_context
-        )._get_real_price_currency(
-            product,
-            rule_id,
-            self.product_uom_qty,
-            self.product_uom,
-            self.folio_id.pricelist_id.id,
-        )
-        if currency_id != self.folio_id.pricelist_id.currency_id.id:
-            base_price = (
-                self.env["res.currency"]
-                .browse(currency_id)
-                .with_context(product_context)
-                .compute(base_price, self.folio_id.pricelist_id.currency_id)
-            )
-        # negative discounts (= surcharge) are included in the display price
-        return max(base_price, final_price)
-
-    @api.multi
-    @api.onchange("product_id")
-    def product_id_change(self):
-        """
-        @param self: object pointer
-        """
-        if not self.product_id:
-            return {"domain": {"product_uom": []}}
-
-        vals = {}
-        domain = {
-            "product_uom": [
-                ("category_id", "=", self.product_id.uom_id.category_id.id)
-            ]
-        }
-        if not self.product_uom or (
-            self.product_id.uom_id.id != self.product_uom.id
-        ):
-            vals["product_uom"] = self.product_id.uom_id
-
-        product = self.product_id.with_context(
-            lang=self.folio_id.partner_id.lang,
-            partner=self.folio_id.partner_id.id,
-            quantity=vals.get("product_uom_qty") or self.product_uom_qty,
-            date=self.folio_id.date_order,
-            pricelist=self.folio_id.pricelist_id.id,
-            uom=self.product_uom.id,
-        )
-
-        result = {"domain": domain}
-
-        title = False
-        message = False
-        warning = {}
-        if product.sale_line_warn != "no-message":
-            title = _("Warning for %s") % product.name
-            message = product.sale_line_warn_msg
-            warning["title"] = title
-            warning["message"] = message
-            result = {"warning": warning}
-            if product.sale_line_warn == "block":
-                self.product_id = False
-                return result
-
-        name = product.name_get()[0][1]
-        if product.description_sale:
-            name += "\n" + product.description_sale
-        vals["name"] = name
-
-        self._compute_tax_id()
-
-        if self.folio_id.pricelist_id and self.folio_id.partner_id:
-            vals["price_unit"] = self.env[
-                "account.tax"
-            ]._fix_tax_included_price_company(
-                self._get_display_price(product),
-                product.taxes_id,
-                self.tax_id,
-                self.company_id,
-            )
-        self.update(vals)
-        return result
-
-    @api.onchange("ser_checkin_date", "ser_checkout_date")
-    def on_change_checkout(self):
-        """
-        When you change checkin_date or checkout_date it will checked it
-        and update the qty of hotel service line
-        -----------------------------------------------------------------
-        @param self: object pointer
-        """
-        if not self.ser_checkin_date:
-            time_a = time.strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-            self.ser_checkin_date = time_a
-        if not self.ser_checkout_date:
-            self.ser_checkout_date = time_a
-        if self.ser_checkout_date < self.ser_checkin_date:
-            raise _("Checkout must be greater or equal checkin date")
-        if self.ser_checkin_date and self.ser_checkout_date:
-            date_a = time.strptime(
-                self.ser_checkout_date, DEFAULT_SERVER_DATETIME_FORMAT
-            )[:5]
-            date_b = time.strptime(
-                self.ser_checkin_date, DEFAULT_SERVER_DATETIME_FORMAT
-            )[:5]
-            diffDate = datetime.datetime(*date_a) - datetime.datetime(*date_b)
-            qty = diffDate.days + 1
-            self.product_uom_qty = qty
-
-    @api.multi
-    def button_confirm(self):
-        """
-        @param self: object pointer
-        """
-        for folio in self:
-            line = folio.service_line_id
-            x = line.button_confirm()
-        return x
-
-    @api.multi
-    def button_done(self):
-        """
-        @param self: object pointer
-        """
-        for folio in self:
-            line = folio.service_line_id
-            x = line.button_done()
-        return x
-
-    @api.multi
-    def copy_data(self, default=None):
-        """
-        @param self: object pointer
-        @param default: dict of default values to be set
-        """
-        sale_line_obj = self.env["sale.order.line"].browse(
-            self.service_line_id.id
-        )
-        return sale_line_obj.copy_data(default=default)
-
-
-class HotelServiceType(models.Model):
-    _name = "hotel.service.type"
-    _description = "Service Type"
-
-    name = fields.Char("Service Name", size=64, required=True)
-    service_id = fields.Many2one("hotel.service.type", "Service Category")
-    child_ids = fields.One2many(
-        "hotel.service.type", "service_id", "Child Categories"
-    )
-
-    @api.multi
-    def name_get(self):
-        def get_names(cat):
-            """ Return the list [cat.name, cat.service_id.name, ...] """
-            res = []
-            while cat:
-                res.append(cat.name)
-                cat = cat.service_id
-            return res
-
-        return [(cat.id, " / ".join(reversed(get_names(cat)))) for cat in self]
-
-    @api.model
-    def name_search(self, name, args=None, operator="ilike", limit=100):
-        if not args:
-            args = []
-        if name:
-            # Be sure name_search is symetric to name_get
-            category_names = name.split(" / ")
-            parents = list(category_names)
-            child = parents.pop()
-            domain = [("name", operator, child)]
-            if parents:
-                names_ids = self.name_search(
-                    " / ".join(parents),
-                    args=args,
-                    operator="ilike",
-                    limit=limit,
-                )
-                category_ids = [name_id[0] for name_id in names_ids]
-                if operator in expression.NEGATIVE_TERM_OPERATORS:
-                    categories = self.search([("id", "not in", category_ids)])
-                    domain = expression.OR(
-                        [[("service_id", "in", categories.ids)], domain]
-                    )
-                else:
-                    domain = expression.AND(
-                        [[("service_id", "in", category_ids)], domain]
-                    )
-                for i in range(1, len(category_names)):
-                    domain = [
-                        [
-                            (
-                                "name",
-                                operator,
-                                " / ".join(category_names[-1 - i :]),
-                            )
-                        ],
-                        domain,
-                    ]
-                    if operator in expression.NEGATIVE_TERM_OPERATORS:
-                        domain = expression.AND(domain)
-                    else:
-                        domain = expression.OR(domain)
-            categories = self.search(
-                expression.AND([domain, args]), limit=limit
-            )
-        else:
-            categories = self.search(args, limit=limit)
-        return categories.name_get()
-
-
-class HotelServices(models.Model):
-    _name = "hotel.services"
-    _description = "Hotel Services and its charges"
-
-    product_id = fields.Many2one(
-        "product.product",
-        "Service_id",
-        required=True,
-        ondelete="cascade",
-        delegate=True,
-    )
-    categ_id = fields.Many2one(
-        "hotel.service.type", string="Service Category", required=True
-    )
-    product_manager = fields.Many2one("res.users", string="Product Manager")
-
-
-class ResCompany(models.Model):
-    _inherit = "res.company"
-
-    additional_hours = fields.Integer(
-        "Additional Hours",
-        help="Provide the min hours value for \
-                                      check in, checkout days, whatever the \
-                                      hours will be provided here based on \
-                                      that extra days will be calculated.",
-    )
-
-
-class AccountInvoice(models.Model):
-    _inherit = "account.invoice"
-
-    @api.model
-    def create(self, vals):
-        res = super(AccountInvoice, self).create(vals)
-        if self._context.get("folio_id"):
-            folio = self.env["hotel.folio"].browse(self._context["folio_id"])
-            folio.write(
-                {"hotel_invoice_id": res.id, "invoice_status": "invoiced"}
-            )
-        return res
